Change default governor to 'ondemand' for RPi 0/0w/1.
On most systems, 'schedutil' is used as default CPU frequency governor,
but that requires CONFIG_SMP which is not enabled in the config
for RPi 0/0w/1 as it only has a single core.
With 'performance', the CPU runs at full speed all the time, which is
unlikely to be desirable for RPi's, so change the default to 'ondemand'.

The default governor needs to have the corresponding kernel
module to be builtin, so change the 'ondemand' one to '=y'.