Add support for Rockchip's RK805 PMIC.
The PINCTRL_RK805 makes the LEDs work on RK3328/RK3228 SOCs based
devices like the Rock64 and can be managed/configured via
/sys/devices/platform/leds/leds/.
The default is as configured in the DeviceTree: mmc0 and heartbeat.
Closes: #990536

With INPUT_RK805_PWRKEY one can invoke various power functions of the
RK805 PMIC via button(s) on the device.