Make CONFIG_MFD_RK808 builtin.
It's already builtin on armhf and it's a dependency of PINCTL_RK805 and
to make that builtin, MFD_RK808 has to be builtin as well.

It also fixes commit fe29debb to fix bug #860976, where
CONFIG_COMMON_CLK_RK808 was meant to be builtin, but checking
/boot/config-5.10.0-8-arm64 reveals it turned out as module because it
too depends on MFD_RK808.
MFD_RK808 kernel module is needed for at least the RK805, RK808, RK809,
RK817 and RK818 PMIC chips from Rockchip.

https://www.cnx-software.com/pdf/Rockchip-Roadmap-2019-.11_mts.pdf
has a diagram on page 5 with Rockchip's PMIC and their usage in SOCs.