[armhf] media: Set MEDIA_SUPPORT=m so drivers are no longer built-in
The top-level config has CONFIG_MEDIA_SUPPORT=m, but the armmp and
armmp-lpae flavours have always overridden it to =y.  That results in
CONFIG_MEDIA_TUNER=y.  Together with CONFIG_MEDIA_SUBDRV_AUTOSELECT=y
this in turns results in the V4L2 and DVB cores and many tuner drivers
being built-in.

This seems to be just a mistake, as there's no need to have these
drivers built in.  Remove the override.