debian/control: Add version to libopencsd-dev build-dependency
When backporting to bullseye, I found that perf can no longer be built
with its version of opencsd.  The minimum version was bumped to 1.0.0
in 5.12 and then again to 1.1.1 in 5.14, but perf would quietly
disable use of opencsd if it was too old.  Starting with 5.15 this is
instead a fatal error.

So put the current minimum version in Build-Depends.