d/rules: Filter out -s option so that Kbuild is not silent
dpkg-buildpackage now passes the -s option to make if
DEB_BUILD_OPTIONS contains terse, but that triggers Kbuild to turn off
progress messages completely.  This doesn't seem to match the intended
meaning of the option, so filter out -s.

It would be preferable to do this in debian/rules.real, so the -s flag
is still in effect for debian/rules and debian/rules.gen, but when I
tried that the -j and --job-server options were lost so the build
became serial.

It's also necessary to use 'override' when replacing MAKEFLAGS, though
not when appending.