d/rules, d/salsa-ci.yml: Enable blhc, ignoring terse output
blhc wants to see complete compiler command lines and complains when
it sees the summary lines that appear with the terse option.  However,
currently the terse option mostly affects the output for builds of
kernel code, which need different hardening options that blhc doesn't
know about.

So add a line to the log when building with terse, telling blhc to
ignore the CC, CCLD, LD, and lINK summary lines.