d/certs: Add certificate and key to enable test signing in CI
In order to test the linux-signed-amd64 source package and its
binaries, we need a non-production certificate and key for signing
code.  These could be generated within the CI pipeline but would
then need to be carried around as artifacts, so it's simpler to
keep them in the source tree.

These were generated with OpenSSL configuration:

    [ req ]
    encrypt_key = no
    x509_extensions = v3_ca
    prompt= no
    distinguished_name = req_distinguished_name

    [ req_distinguished_name ]
    CN = Test Signer CA

    [ v3_ca ]
    subjectKeyIdentifier = hash
    authorityKeyIdentifier = keyid:always,issuer
    basicConstraints = critical,CA:true

and command line:

    openssl req -new -x509 -newkey rsa:2048 \
        -keyout debian/certs/ci-test-sign/ci-test-sign-key.pem \
	-out debian/certs/ci-test-sign/ci-test-sign.pem