Update to 5.18rc6
When more then a simple rebase is done, a relevant upstream commit ID is
mentioned between brackets.

The debian/android-enable-building-ashmem-and-binder-as-modules.patch
has been partially 'reverted' as upstream removed support for ashmem
entirely.

The debian/module-avoid-abi-changes-when-debug-info-is-disabled.patch
has been substantially changed and was done to 'fix' salsa's CI and I
don't know how to handle that.