udeb: Add essiv to crypto-modules (Closes: #973378)
Since Linux 5.4, the ESSIV algorithm is implemented as a module in the
crypto subsystem, rather than built into dm-crypt.  DM_CRYPT selects
CRYPTO_ESSIV so the new module was included in linux-image packages,
but there is no symbol dependency that would pull essiv into the
installer.

This didn't affect creation of new encrypted volumes in the installer,
because ESSIV isn't used in the current encryption configuration, but
it does prevent reusing older encrypted volumes there.

Add essiv to crypto-modules, and also explicitly enable CRYPTO_ESSIV
in the config in case DM_CRYPT ever stops selecting it.