[hppa,mips,mipsel,powerpc] lintian: Override error for 64-bit kernels
Some 32-bit architectures have 64-bit kernel flavours and we don't
yet mandate use of multiarch for them.  Currently lintian reports a
binary-from-other-architecture error for each of their kernel modules,
and (on MIPS) for vmlinux.

This also automatically adds a lintian-overrides file for signed
linux-image packages.  There's no overlap between the above
architectures and those for which we do code signing, but we might
need overrides from them in future and it would take extra work to
exclude this.  Therefore, add dh-exec to build-dependencies of the
linux-signed source packages.