d/tests/kbuild: Fix default-flavour lookup for arches with no featuresets
We currently look for a default-flavour definition for the given arch
and featureset=none.  But for architectures that have no featuresets,
it is currently defined at the arch level and we don't find it.

getconfig.py probably should provide the option to do a "merged"
lookup that will fall-back to a more general definition, as used in
gencontrol.  For now, implement the fallback here.

This affected release architectures ppc64el and s390x, plus non-
release architectures alpha, hppa, ia64, m68k, ppc64, and sparc64.