udeb: Move ledtrig-audio from sound-modules to kernel-image
ledtrig-audio isn't explicitly included in any package, but several
drivers in sound-modules depend on it so it's included there.

Now in Linux 6.0, asus-wmi also depends on ledtrig-audio (and is
itself a dependency of hid-asus) so ledtrig-audio is included in
input-modules too.  kernel-wedge detects this duplication and fails.

Move ledtrig-audio in kernel-image, since led-class is already there
and it's a trivial amount of extra code.