udeb: Also add drivers in subdirectories of drivers/net/phy
Currently the mscc PHY driver resides in its own subirectory. As we
only include drivers in the toplevel directory it's missing in the
udeb.

Devices such as the Kontron BL i.MX8MM require this driver for
bringing up the NIC, so enable it by extending the wildcard to
include subdirectories.

Signed-off-by: Frieder Schrempf <frieder.schrempf@kontron.de>