d/t/control.tools-unversioned.in: version rtla build deps
Linux commit fe4d0d5dde ("rtla/Makefile: Properly handle dependencies")
introduces stricter checks on available versions of development packages
for libtracefs (>= 1.3) and libtraceevent (>= 1.5).

For Debian unstable, this is not a problem, because we have new enough
versions in there now. But, if we want to do bullseye-backports for
Linux 5.19, the build will FTBFS with the following messages:

********************************************
** NOTICE: libtraceevent version 1.5 or higher not found
**
** Consider installing the latest libtraceevent from your
** distribution, e.g., 'dnf install libtraceevent' on Fedora,
** or from source:
**
**  https://git.kernel.org/pub/scm/libs/libtrace/libtraceevent.git/
**
********************************************

********************************************
** NOTICE: libtracefs version 1.3 or higher not found
**
** Consider installing the latest libtracefs from your
** distribution, e.g., 'dnf install libtracefs' on Fedora,
** or from source:
**
**  https://git.kernel.org/pub/scm/libs/libtrace/libtracefs.git/
**
********************************************

Let's at least version the build dependencies for rtla to make this more
explicit. When trying to build on Bullseye, it will fail immediately,
instead of halfway the whole package build process.

Note: currently, the rtla binary package that is being built will still
display libtraceevent1 (>= 5.4) as minimal dependency, but libtracefs1
(>= 1.3.0) does have a dependency on the actual newer required
libtraceevent1 by itself. Therefore we don't have to introduce another
customization in the Depends field.

[benh: Update d/changelog]