[arm64] Drop "arm64: dts: rockchip: correct voltage selector on Firefly-RK3399"
This patch was applied (in 2018 with kernel 4.17) in anticipation of it
being merged upstream. But that didn't happen (technically: it was but
then dropped again). See https://bugs.debian.org/900799#24

It seems better to align with upstream again and drop this patch.
The patch might alter behavior inconsistent with upstream and that
would be hard to diagnose.
If it turns out that dropping the patch would actually re-introduce the
issue this patch was meant to solve, we would get/have a new bug report
against a recent kernel, which will surely help in getting it actually
fixed upstream.