debian/config: Enable HIST_TRIGGERS for all kernels
The HIST_TRIGGERS feature enables the collection of histogram for
traced events based on groupings by key-values. It is a low overhead
mechanism to get aggregate information of system behaviour. E.g., It
can be used to get an histogram of the number and types of syscalls
being executed in the system.

As it is generically useful, enable the config for all builds.