d/templates: Improve package description for "header" packages
The package containing the architecture-specific kernel header files
mentioned the installation location inside its extended package
description, whereas the package containing the common kernel header
files and the package containing the header files for userspace did not.

This commit adds the installation location to the package containing the
common kernel header files and to the package that contains the header
files for userspace.

[benh: Add changelog line]