d/b/test-patches: Change ABI name to make packages co-installable
One of the problems raised in bug #1022061 is that the linux-image
packages built by test-patches conflict with official packages.  They
should be co-installable so that it's easy to switch back to running
the official kernel.

Secondly, some patches might change ABI.  Depending on the state of
the source before running test-patches, this can result in a build
failure.

For both these reasons, override the ABI name.  Do this through the
newly supported debian/config.local directory so it's easily
reversible.  Set the ABI name to "0.a.test" so that any official
package (including backports) will be ordered before it in boot menus.