linux-perf: Add libtraceevent-dev to Build-Depends
Since Linux 6.2, perf is not build with an in-tree copy of
libtraceevent, so we should Build-Depend on libtraceevent-dev.

On architectures where we build rtla, we already have this
build-dependency.  On other architectures, perf's Makefile.config
initially warned but continued without libtraceevent, resulting in a
functional regression.  In 6.4 this was changed to be a fatal error.

This caused FTBFS on many architectures - probably alpha, armel, hppa,
mips*, powerpc*, riscv64, s390x, sh4, and sparc64.