Add pkg.linux.mintools profile for building minimal userland tools
debian/bin/test-patches needs to build the linux-kbuild-<version>
and (on powerpc and ppc64) linux-bootwrapper-<abiname> packages to
satisfy the dependencies of the packages it currently builds.

This could be done by dropping the pkg.linux.notools profile, but that
would require users to install many extra build-dependencies.

Instead, add a new profile which builds these tools and no others.