[x86] Enable modules for more ChromeOS hardware functionality
The ChromeOS Embedded Controller driver was enabled as a module a while
back, but unfortunately this doesn't automatically enable the various
ways (I2C, SPI, LPC, ISHTP) to talk to the EC. Enable these. Also enable
drivers for various sensors and other functionality that the EC exposes
depending on hardware.

There seems to be another "Wilco EC" on some Dell Chromebooks, enable
drivers for it as well. Also enable drivers for the privacy screen and a
tablet switch controller on some Chromebooks.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>