[x86] Enable more SoC audio modules for Intel and AMD systems
Enable various sound-related modules with the intention to improve audio
support on most x86 Chromebooks. Looks like SOF support is desirable for
ChromeOS devices using Broadwell and Bay Trail, but this does not enable
it as those were disabled per bug #945914. And there are reports of
hardware damage due to MAX98357A on AVS drivers, so explicitly disable
that with a note.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>