[arm64] Add qrtr to kernel-image udeb
Netboot installer images work fine on the Thinkpad X13s, while the netinst
versions boot into a blank display. This is due to qrtr.ko not being included
in the netinst initrd. The module is instead included in netboot images by
virtue of being a dependency of qrtr_mhi, shipped with the nic-wireless-modules
udeb.

Explicitly add qrtr to kernel-image for arm64.