d/patches: Remove amdgpu part in 2 firmware patches
According to bug 1053764 the amdgpu driver has been modified so that it
now checks itself whether the needed firmware is available and loads
that. Or if it isn't, then it will continue to use the framebuffer
provided by the boot firmware.

IOW: The functionality of this patch has been incorporated upstream for
the amdgpu (but not for the radeon) driver.

Link: https://bugs.debian.org/1053764