d/rules.real: Run dh_movetousr if available
dh_movetousr will move files from /lib to /usr/lib in trixie and later
releases, and will be a no-op when it is added to bookworm-backports.
To support backports to older suites, check whether it is available
before running it.

This specifically moves:
- The udev rules file in hyperv-daemons
- All the modules and module metadata in linux-image-*
- The build and source symlinks in linux-headers-*

The systemd units in hyperv-daemons are already moved as the result of
a change in dh_installsystemd.